Pull out dumping/loading external modules into separate module
Also split apart dumping/loading into separate functions.

This all helps resolve potential circular imports in future
